+++ test-java-lib/src/main/java/com/uber/lib/unannotated/RestrictivelyAnnotatedClass.java
@@ -1,0 +1,25 @@
+package com.uber.lib.unannotated;
+
+import org.checkerframework.checker.nullness.qual.NonNull;
+import org.checkerframework.checker.nullness.qual.Nullable;
+
+public class RestrictivelyAnnotatedClass {
+
+  private final Object field;
+
+  public RestrictivelyAnnotatedClass(Object field) {
+    this.field = field;
+  }
+
+  public @Nullable Object getField() {
+    return field;
+  }
+
+  public static @Nullable Object returnsNull() {
+    return null;
+  }
+
+  public static void consumesObjectNonNull(@NonNull Object o) {}
+
+  public static void consumesObjectUnannotated(Object o) {}
+}